requirements.txt
eventlet
jinja2
pigwig
pillow
pyyaml
